+++ libs/shinkai-tools-runner/src/tools/python_runner_options.rs
@@ -6,25 +6,19 @@ use super::{
 };
 
 #[derive(Clone)]
-pub struct DenoRunnerOptions {
+pub struct PythonRunnerOptions {
     pub context: ExecutionContext,
-    pub deno_binary_path: PathBuf,
     pub python_binary_path: PathBuf,
     pub code_runner_docker_image_name: String,
     pub force_runner_type: Option<RunnerType>,
     pub shinkai_node_location: ShinkaiNodeLocation,
 }
 
-impl Default for DenoRunnerOptions {
+impl Default for PythonRunnerOptions {
     fn default() -> Self {
         Self {
             context: ExecutionContext::default(),
             code_runner_docker_image_name: String::from("dcspark/shinkai-code-runner:0.9.0"),
-            deno_binary_path: PathBuf::from(if cfg!(windows) {
-                "./shinkai-tools-runner-resources/deno.exe"
-            } else {
-                "./shinkai-tools-runner-resources/deno"
-            }),
             python_binary_path: PathBuf::from(if cfg!(windows) {
                 "C://Users/agall/AppData/Local/Microsoft/WindowsApps/python"
             } else {
