+++ libs/shinkai-tools-runner/src/tools/runner_type.rs
@@ -1,0 +1,5 @@
+#[derive(Clone)]
+pub enum RunnerType {
+    Host,
+    Docker,
+}
