+++ libs/shinkai-tools-runner/src/tools/shinkai_node_location.rs
@@ -1,0 +1,6 @@
+#[derive(Clone)]
+pub struct ShinkaiNodeLocation {
+    pub protocol: String,
+    pub host: String,
+    pub port: u16,
+}
